feat:add gray-release demo (#92)
