--- polaris-agent-examples/spring-cloud-plugins-examples/spring-cloud-2021-examples/consumer/src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.application.name=service-consumer
-server.port=18083
-spring.cloud.nacos.discovery.server-addr=127.0.0.1:8848
-spring.cloud.nacos.username=nacos
-spring.cloud.nacos.password=nacos
