+++ polaris-agent-examples/spring-cloud-plugins-examples/spring-cloud-2021-examples/gray-release-examples/pom.xml
@@ -3,20 +3,20 @@
 		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<parent>
-		<artifactId>spring-cloud-plugins-examples</artifactId>
+		<artifactId>spring-cloud-2021-examples</artifactId>
 		<groupId>com.tencent.polaris</groupId>
 		<version>${revision}</version>
 		<relativePath>../pom.xml</relativePath>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
 
-	<artifactId>spring-cloud-2021-examples</artifactId>
-
+	<artifactId>gray-release-examples</artifactId>
 	<packaging>pom</packaging>
-
 	<modules>
-		<module>gray-release-examples</module>
-		<module>quickstart-examples</module>
+		<module>service-a</module>
+		<module>service-b</module>
+		<module>service-c</module>
 	</modules>
+
 
 </project>
