+++ polaris-agent-examples/spring-cloud-plugins-examples/spring-cloud-2021-examples/gray-release-examples/service-c/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.application.name=service-c
+server.port=0
+spring.cloud.nacos.discovery.server-addr=127.0.0.1:8848
+spring.cloud.nacos.username=nacos
+spring.cloud.nacos.password=nacos
