+++ polaris-agent-examples/spring-cloud-plugins-examples/spring-cloud-2021-examples/quickstart-examples/provider/src/main/java/cn/polarismesh/agent/examples/alibaba/cloud/cloud/ProviderApplication.java
@@ -1,0 +1,59 @@
+/*
+ * Tencent is pleased to support the open source community by making Polaris available.
+ *
+ * Copyright (C) 2019 THL A29 Limited, a Tencent company. All rights reserved.
+ *
+ * Licensed under the BSD 3-Clause License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed
+ * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ * CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package cn.polarismesh.agent.examples.alibaba.cloud.cloud;
+
+import com.alibaba.nacos.common.utils.JacksonUtils;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.serviceregistry.Registration;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
+ */
+@SpringBootApplication
+public class ProviderApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ProviderApplication.class, args);
+	}
+
+	@RestController
+	public static class EchoController {
+
+		private Registration registration;
+
+		public EchoController(Registration registration) {
+			this.registration = registration;
+		}
+
+		@GetMapping("/echo/{string}")
+		public String echo(@PathVariable String string) {
+			String sb = "Hello, I'm provider, receive msg : "
+					+ string
+					+ "my metadata : "
+					+ JacksonUtils.toJson(registration.getMetadata());
+			return sb;
+		}
+
+	}
+
+}
