+++ polaris-agent-plugins/spring-cloud-plugins/spring-cloud-2021-plugin/src/main/resources/default-plugin.conf
@@ -1,50 +1,57 @@
+# 配置北极星服务端地址
 spring.cloud.polaris.address=grpc\://127.0.0.1\:8091
 spring.cloud.polaris.enabled=true
 
-## ??????
+# 启用 Java Agent 的 Spring Cloud Tencent 注册发现能力
+spring.cloud.discovery.enabled=true
+# 配置服务注册发现的命名空间信息
 spring.cloud.polaris.discovery.namespace=default
+# 启用从北极星
 spring.cloud.polaris.discovery.enabled=true
 spring.cloud.polaris.discovery.register=true
 
-## ????
+## 是否启用北极星的 LoadBalancer
 spring.cloud.polaris.loadbalancer.enabled=true
 #spring.cloud.polaris.loadbalancer.discoveryType=POLARIS
 #spring.cloud.polaris.loadbalancer.strategy=random
 
-## ????
-spring.cloud.polaris.router.enabled=true
-# ?????
-spring.cloud.polaris.router.rule-router.enabled=true
-# ?????
-spring.cloud.polaris.router.metadata-router.enabled=true
-# ????
-spring.cloud.polaris.router.nearby-router.enabled=true
+## 是否启用北极星服务路由能力
+spring.cloud.polaris.router.enabled=false
+# 是否启用北极星的自定义路由能力
+spring.cloud.polaris.router.rule-router.enabled=false
+# 是否启用北极星的元数据路由能力
+spring.cloud.polaris.router.metadata-router.enabled=false
+# 是否启用北极星的就近路由能力
+spring.cloud.polaris.router.nearby-router.enabled=false
 
-## ????
+## 是否启用北极星的服务限流能力
 spring.cloud.polaris.ratelimit.enabled=true
-#spring.cloud.polaris.ratelimit.rejectRequestTips=""
+# 设置触发限流时的提示信息
+# spring.cloud.polaris.ratelimit.rejectRequestTips=""
+# 设置触发限流时的响应码
 spring.cloud.polaris.ratelimit.rejectHttpCode=429
-# ?????????????????
+# 设置限流匀速排队最大排队时间
 spring.cloud.polaris.ratelimit.maxQueuingTime=1000
 
 
-## RPC ????
+## RPC 调用增强
 spring.cloud.tencent.rpc-enhancement.enabled=true
-# ??????
+# 开启 RPC 调用结果上报
 spring.cloud.tencent.rpc-enhancement.reporter.enabled=true
-#
-#spring.cloud.tencent.rpc-enhancement.reporter.ignore-internal-server-error=true
-#spring.cloud.tencent.rpc-enhancement.reporter.series=
-#spring.cloud.tencent.rpc-enhancement.reporter.statuses=""
 
-#spring.cloud.polaris.stat.host=""
+# 配置北极星监控指标上报
 spring.cloud.polaris.stat.enabled=true
+# 指标上报监听端口
 spring.cloud.polaris.stat.port=0
+# 指标上报暴露的 http path
 spring.cloud.polaris.stat.path=/metrics
 
-## ????
+## 启用北极星配置中心能力
 spring.cloud.polaris.config.enabled=false
+# 配置北极星配置中心地址
 spring.cloud.polaris.config.address=grpc\://127.0.0.1\:8093
+# 是否开启配置监听刷新
 spring.cloud.polaris.config.auto-refresh=true
+# Spring Cloud Tencent 配置刷新机制
 spring.cloud.polaris.config.refresh-type=refresh_context
